Under FreeBSD, Binutils objcopy always put the section header at the end of all the sections. We want to do the same here.
The purpose is that after this change we are now able to generate
identical binaries with Binutils objcopy/strip on FreeBSD . This will
make debugging elfcopy much easier since we can diff the binary if we
notice something is wrong.

However, note that the behaviour is still different with Binutils:
elfcopy checks the FreeBSD OSABI tag to tell whether it needs to
move the section headers, while Binutils is probably configured
this way when it's compiled on FreeBSD.

Also, I think the testsuite for elfcopy need a complete regen after
this change.

git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@3213 95820547-d848-0410-985e-9ae8fe0fa350